squash a -Wsign-conversion warning in iofread
The preceding call to fgets can only have returned at most bufsize - 1
characters, so it is safe to downcast the size_t result of strlen to an int.

+++ lib/cgraph/io.c
@@ -17,7 +17,7 @@
 static int iofread(void *chan, char *buf, int bufsize)
 {
     if (fgets(buf, bufsize, chan))
-	return strlen(buf);
+	return (int)strlen(buf);
     else
 	return 0;
 }
